Avanços na camada de repository
