+++ app/src/main/java/com/davidmag/movienatic/rest/ResourceCallResult.java
@@ -1,0 +1,5 @@
+package com.davidmag.movienatic.rest;
+
+public class ResourceCallResult<T> {
+
+}
